move types out of experiment-form-def.tsx into experiment-form-types.ts
