+++ src/app/experiments/create/experiment-form/experiment-form-def.tsx
@@ -5,16 +5,7 @@ import {
   ExperimentMetadataScreen,
 } from '@/app/experiments/create/experiment-form/experiment-metadata-screen';
 import { ExperimentTypeScreen } from '@/app/experiments/create/experiment-form/experiment-type-screen';
-import {
-  Arm,
-  ContextType,
-  CreateExperimentResponse,
-  DesignSpecInput,
-  FieldMetadata,
-  FilterInput,
-  GetFiltersResponseElement,
-  PowerResponseOutput,
-} from '@/api/methods.schemas';
+import { ContextType } from '@/api/methods.schemas';
 import { abandonExperiment } from '@/api/admin';
 import { ExperimentSelectDatasourceScreen } from '@/app/experiments/create/experiment-form/experiment-select-datasource-screen';
 import { ExperimentSelectBinaryOrRealOutcomes } from '@/app/experiments/create/experiment-form/experiment-select-binary-or-real-outcomes';
@@ -41,64 +32,12 @@ import {
   toCmabBanditParams,
   toMabBanditParams,
 } from '@/app/experiments/create/experiment-form/experiment-bandit-helpers';
-import { ErrorType } from '@/services/orval-fetch';
 import {
-  BanditParams,
+  ExperimentFormData,
+  ExperimentScreenId,
+  ExperimentType,
   isBanditExperimentType,
-  MetricWithMDE,
-  PowerCheckOption,
 } from '@/app/experiments/create/experiment-form/experiment-form-types';
-
-export type ExperimentType = DesignSpecInput['experiment_type'];
-
-// Defines the entirety of the editable data collected via this wizard flow.
-export type ExperimentFormData = {
-  // experiment-metadata-screen
-  name?: string;
-  hypothesis?: string;
-  designUrl?: string;
-  startDate?: string;
-  endDate?: string;
-
-  // experiment-type-screen
-  experimentType?: ExperimentType;
-
-  // experiment-select-datasource-screen
-  datasourceId?: string;
-  tableName?: string;
-
-  // experiment-freq-stack-screen
-  primaryKey?: string;
-  primaryMetric?: MetricWithMDE;
-  secondaryMetrics?: MetricWithMDE[];
-  filters?: FilterInput[];
-  // Cache of available filter fields (and their data types) for lookup/display/search
-  availableFilterFields?: GetFiltersResponseElement[];
-  strata?: FieldMetadata[];
-  // These next 2 Experiment Parameters are strings to allow for empty values,
-  // which should be converted to numbers when making power or experiment creation requests.
-  confidence?: string;
-  power?: string;
-  // Populated when user clicks "Power Check" on DesignForm
-  desiredN?: number;
-  sampleSizeOption?: PowerCheckOption;
-  powerCheckResponse?: PowerResponseOutput;
-  createExperimentResponse?: CreateExperimentResponse;
-  createExperimentError?: ErrorType<unknown>;
-
-  // experiment-describe-webhooks-screen
-  selectedWebhookIds?: string[];
-
-  // experiment-describe-arms-screen
-  arms?: Omit<Arm, 'arm_id'>[];
-
-  // bandit flow config
-  bandit?: BanditParams;
-
-  // experiment-summarize-freq-screen (populated after createExperiment API call)
-  experimentId?: string;
-  commitError?: ErrorType<unknown>;
-};
 
 const isFreq = (experimentType: ExperimentType) => {
   return experimentType === 'freq_online' || experimentType === 'freq_preassigned';
@@ -107,20 +46,6 @@ const isFreq = (experimentType: ExperimentType) => {
 const isCmab = (experimentType: ExperimentType) => {
   return experimentType === 'cmab_online';
 };
-
-// Defines a type for all known screen IDs for the experiment form. This type is used with screen() to
-// type-check the ids returned by nextScreen and prevScreen.
-export type ExperimentScreenId =
-  | 'metadata'
-  | 'experiment-type'
-  | 'freq-select-datasource'
-  | 'bandit-binary-or-real'
-  | 'describe-contexts'
-  | 'describe-arms'
-  | 'describe-bandit-arms'
-  | 'freq-stack'
-  | 'summarize-freq'
-  | 'summarize-bandit';
 
 // Helper to create screens with proper type inference
 const screen = packScreen<ExperimentFormData, ExperimentScreenId>();
